fix(a11y): optimizations and more intuitive file locations
